+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/utils/generate-user-pool-group-stack-template.ts
@@ -2,6 +2,9 @@ import { $TSContext } from 'amplify-cli-core';
 import { Template } from 'cloudform-types';
 import { AmplifyUserPoolGroupTransform } from '../auth-stack-builder/user-pool-group-stack-transform';
 
+/**
+ * Entry Point to generate UserPoolGroup stack template
+ */
 export const generateUserPoolGroupStackTemplate = async (context: $TSContext, resourceName: string): Promise<Template> => {
   try {
     const userPoolTransform = new AmplifyUserPoolGroupTransform(resourceName);