+++ packages/amplify-e2e-tests/src/__tests__/function_5.test.ts
@@ -14,6 +14,7 @@ import {
   updateFunction,
   amplifyPushWithoutCodegen,
   addFeatureFlag,
+  generateRandomShortId,
 } from 'amplify-e2e-core';
 import _ from 'lodash';
 
@@ -34,8 +35,7 @@ describe('test initEnv() behavior in function', () => {
 
   it('init a project and add simple function and uncomment cors header', async () => {
     await initJSProjectWithProfile(projRoot, { disableAmplifyAppCreation: false });
-    const random = Math.floor(Math.random() * 10000);
-    const functionName = `testfunction${random}`;
+    const functionName = `testfunction${generateRandomShortId()}`;
     await addFunction(projRoot, { functionTemplate: 'Hello World', name: functionName }, 'nodejs');
     await amplifyPushAuth(projRoot);
     const meta = getProjectMeta(projRoot);
@@ -82,7 +82,7 @@ describe('test dependency in root stack', () => {
     deleteProjectDir(projRoot);
   });
 
-  it('init a project with api and function and update the @model and add function access to @model ', async () => {
+  it('init a project with api and function and update the @model and add function access to @model', async () => {
     const projectName = 'mytestapi';
     await initJSProjectWithProfile(projRoot, {
       name: projectName,
@@ -90,8 +90,7 @@ describe('test dependency in root stack', () => {
     await addApiWithoutSchema(projRoot, { transformerVersion: 1 });
     await updateApiSchema(projRoot, projectName, 'simple_model.graphql');
 
-    const random = Math.floor(Math.random() * 10000);
-    const fnName = `integtestfn${random}`;
+    const fnName = `integtestfn${generateRandomShortId()}`;
     await addFunction(
       projRoot,
       {