+++ packages/amplify-provider-awscloudformation/src/utility-functions.js
@@ -275,7 +275,7 @@ module.exports = {
    * @deprecated Use getGraphQLAPIs instead
    */
   getAppSyncAPIs: context => {
-    return this.getGraphQLAPIs(context);
+    return module.exports.getGraphQLAPIs(context);
   },
   getGraphQLAPIs: context => {
     const log = logger('getGraphQLAPIs.appSyncModel.appSync.listGraphqlApis', { maxResults: 25 });
@@ -374,7 +374,7 @@ module.exports = {
    * @deprecated Use getGraphQLApiKeys instead
    */
   getAppSyncApiKeys: (context, options) => {
-    return this.getGraphQLApiKeys(context, options);
+    return module.exports.getGraphQLApiKeys(context, options);
   },
   getGraphQLApiKeys: (context, options) => {
     const awsOptions = {};