+++ packages/amplify-java-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.30](https://github.com/aws-amplify/amplify-cli/compare/amplify-java-function-runtime-provider@2.2.29...amplify-java-function-runtime-provider@2.2.30) (2022-05-24)
+
+**Note:** Version bump only for package amplify-java-function-runtime-provider
+
+
+
+
+
 ## [2.2.29](https://github.com/aws-amplify/amplify-cli/compare/amplify-java-function-runtime-provider@2.2.28...amplify-java-function-runtime-provider@2.2.29) (2022-05-10)
 
 **Note:** Version bump only for package amplify-java-function-runtime-provider